chore(servers): move components into folders and improve naming
